Move API under Specification submodule
